requirements.txt
python-twitter
praw
configparser
